Update test infraestructure to enable re-running tests
