+++ demo_widget.html
@@ -10,6 +10,5 @@
       data-app="YesGraph" data-email="some@email.com"
       data-name="Some Name" data-foo="bar" data-options-id='staff'>
 </span>
-
 </body>
 </html>
