+++ tests/master.js
@@ -1,0 +1,5 @@
+var testApi = require("./specs/test_api.js");
+var testSuperwidget = require("./specs/test_superwidget.js");
+
+testApi();
+testSuperwidget();
